Move shared code to apollo-shared package
