+++ packages/apollo-shared/package.json
@@ -6,6 +6,10 @@
     "build": "tsc"
   },
   "dependencies": {
+    "@gmod/gff": "^1.2.0",
+    "@jbrowse/core": "^1.5.9",
+    "mobx-state-tree": "3.14.1",
+    "regenerator-runtime": "^0.13.9",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
