+++ packages/apollo-shared/src/BackendDrivers/index.ts
@@ -1,0 +1,3 @@
+export * from './AnnotationFeature'
+export * from './BackendDriver'
+export * from './CollaborationServerDriver'
