+++ packages/apollo-shared/src/ChangeManager/ChangeManager.ts
