+++ packages/apollo-shared/src/ChangeManager/ChangeTypes.ts
