+++ packages/apollo-shared/src/ChangeManager/index.ts
@@ -1,0 +1,4 @@
+export * from './Change'
+export * from './ChangeManager'
+export * from './ChangeTypes'
+export * from './LocationEndChange'
