+++ packages/apollo-shared/src/index.ts
@@ -1,1 +1,3 @@
-export const apolloShared = 'apollo-shared'
+export * from './BackendDrivers'
+export * from './ChangeManager'
+export * from './Validations'
