+++ packages/jbrowse-plugin-apollo/package.json
@@ -50,6 +50,7 @@
   "dependencies": {
     "@gmod/gff": "^1.2.0",
     "@jbrowse/plugin-linear-genome-view": "^1.5.1",
+    "apollo-shared": "workspace:^",
     "regenerator-runtime": "^0.13.9",
     "tslib": "^2.0.3"
   },
